better way to return a double

+++ GroupProject/JobCompare6300/app/src/main/java/edu/gatech/seclass/jobcompare6300/ComparisonActivity.java
@@ -171,9 +171,9 @@ public class ComparisonActivity extends AppCompatActivity {
         }
 
         double jobScore = 0;
-        int weightTotal = bonus + salary + commute + leave+ retirement;
-        jobScore = Double.valueOf(((salary/weightTotal) * AYS) + ((bonus/weightTotal) * AYB) + ( (retirement/weightTotal) * (RBP * AYS))
-                + ( (leave/weightTotal) * (LT * AYS / 260) ) - ( (commute/weightTotal) * (CT * AYS / 8) ));
+        double weightTotal = bonus + salary + commute + leave+ retirement;
+        jobScore = ((salary/weightTotal) * AYS) + ((bonus/weightTotal) * AYB) + ( (retirement/weightTotal) * (RBP * AYS))
+                + ( (leave/weightTotal) * (LT * AYS / 260) ) - ((commute/weightTotal) * (CT * AYS / 8));
         return jobScore;
     }
 }
